Black formatting

+++ hrpe/features/deltas.py
@@ -5,7 +5,7 @@ def calculate_deltas(data):
     """
 	input
 	data - hh data, use transform.
-    
+
     return hh_data with extra columns:
 	delta max - maximum value over the 30 minute perod subtracted by the mean.
 	delta min - mean value over the 30 minute perod subtracted by the minimum value.
@@ -19,7 +19,7 @@ def calculate_deltas(data):
     # Aggregrate to max min and mean
     hh_data["delta_max"] = hh_data["value_max"] - hh_data["value_mean"]
     hh_data["delta_min"] = hh_data["value_mean"] - hh_data["value_min"]
-    hh_data['range'] = hh_data["value_max"] - hh_data["value_min"]
+    hh_data["range"] = hh_data["value_max"] - hh_data["value_min"]
 
     return hh_data
 
@@ -28,7 +28,7 @@ def calculate_deltas(data):
     """
 	input
 	data - hh data, use transform.
-    
+
     return hh_data with extra columns:
 	delta max - maximum value over the 30 minute perod subtracted by the mean.
 	delta min - mean value over the 30 minute perod subtracted by the minimum value.
@@ -42,11 +42,12 @@ def calculate_deltas(data):
     # Aggregrate to max min and mean
     hh_data["delta_max"] = hh_data["value_max"] - hh_data["value_mean"]
     hh_data["delta_min"] = hh_data["value_mean"] - hh_data["value_min"]
-    hh_data['range'] = hh_data["value_max"] - hh_data["value_min"]
+    hh_data["range"] = hh_data["value_max"] - hh_data["value_min"]
 
     return hh_data
 
- 	hh_data["is_weekday"] = hh_data["is_weekday"].astype(int)
+
+# hh_data["is_weekday"] = hh_data["is_weekday"].astype(int)
 
 
 def calculate_maxmin(data):
